+++ src/controller/index.ts
@@ -1,0 +1,2 @@
+export { findUserController } from './user';
+export { findNoteController } from './note'
